STL-716: [database] Add transaction to database service
