+++ database/pubspec.yaml
@@ -1,6 +1,6 @@
 name: database_service
 description: Database service module
-version: 2.4.0
+version: 2.5.0
 publish_to: none
 
 environment:
@@ -9,6 +9,7 @@ environment:
 dependencies:
   async: ^2.13.0
   collection: ^1.19.1
+  meta: ^1.17.0
   path: ^1.9.1
   sqflite: ^2.4.2
